Move compatible `:navigation:navigation-common` unit tests to common.
Testing: `./gradlew :navigation:navigation-common:desktopTest`
