+++ navigation/navigation-common/build.gradle
@@ -39,9 +39,6 @@ plugins {
 
 JetbrainsAndroidXPlugin.applyAndConfigure(project)
 
-def jsEnabled = KmpPlatformsKt.enableJs(project)
-def wasmEnabled = KmpPlatformsKt.enableWasm(project)
-
 androidXMultiplatform {
     android()
     desktop()
@@ -91,12 +88,12 @@ kotlin {
                 implementation(libs.kotlinCoroutinesTest)
                 implementation(libs.kotlinTest)
                 implementation(project(":kruth:kruth"))
+                implementation(project(":navigation:navigation-testing"))
             }
         }
 
-        jvmMain {
-            dependsOn(commonMain)
-        }
+        jvmMain.dependsOn(commonMain)
+        jvmTest.dependsOn(commonTest)
 
         androidMain {
             dependsOn(jvmMain)
@@ -107,10 +104,8 @@ kotlin {
         }
 
         androidUnitTest {
-            dependsOn(commonTest)
+            dependsOn(jvmTest)
             dependencies {
-                // TODO: uncomment once KMP for navigation-runtime is ready
-                //implementation(project(":navigation:navigation-testing"))
                 implementation("androidx.arch.core:core-testing:2.2.0")
                 implementation(libs.junit)
                 implementation(libs.mockitoCore4)
@@ -119,7 +114,6 @@ kotlin {
         }
 
         androidInstrumentedTest {
-            dependsOn(commonTest)
             dependencies {
                 implementation(libs.kotlinTestJunit)
                 implementation(libs.testExtJunit)
@@ -144,38 +138,34 @@ kotlin {
             dependsOn(jvmMain)
             dependsOn(jbMain)
         }
-        desktopTest.dependsOn(jbTest)
+        desktopTest {
+            dependsOn(jvmTest)
+            dependsOn(jbTest)
+        }
         nonJvmMain.dependsOn(jbMain)
         nativeMain.dependsOn(nonJvmMain)
-        nativeTest.dependsOn(jbTest)
+        webMain.dependsOn(nonJvmMain)
         darwinMain.dependsOn(nativeMain)
         linuxMain.dependsOn(nativeMain)
 
-        if (jsEnabled || wasmEnabled) {
-            webMain.dependsOn(nonJvmMain)
-            jsTest.dependsOn(jbTest)
-            wasmJsTest.dependsOn(jbTest)
-        }
+        nonJvmTest.dependsOn(jbTest)
+        nativeTest.dependsOn(nonJvmTest)
+        webTest.dependsOn(nonJvmTest)
+        darwinTest.dependsOn(nativeTest)
+        linuxTest.dependsOn(nativeTest)
 
         targets.all { target ->
             if (target.platformType == KotlinPlatformType.native) {
-                target.compilations["main"].defaultSourceSet {
-                    def konanTargetFamily = target.konanTarget.family
-                    if (konanTargetFamily.appleFamily) {
-                        dependsOn(darwinMain)
-                    } else if (konanTargetFamily == Family.LINUX) {
-                        dependsOn(linuxMain)
-                    } else {
-                        throw new GradleException("unknown native target ${target}")
-                    }
+                if (target.konanTarget.family.appleFamily) {
+                    target.compilations["main"].defaultSourceSet.dependsOn(darwinMain)
+                    target.compilations["test"].defaultSourceSet.dependsOn(darwinTest)
+                } else if (target.konanTarget.family == Family.LINUX) {
+                    target.compilations["main"].defaultSourceSet.dependsOn(linuxMain)
+                    target.compilations["test"].defaultSourceSet.dependsOn(linuxTest)
                 }
-            } else if (target.platformType in [
-                    KotlinPlatformType.js,
-                    KotlinPlatformType.wasm
-            ]) {
-                target.compilations["main"].defaultSourceSet {
-                    dependsOn(webMain)
-                }
+            } else if (target.platformType in [KotlinPlatformType.js, KotlinPlatformType.wasm]) {
+                target.compilations["main"].defaultSourceSet.dependsOn(webMain)
+                target.compilations["test"].defaultSourceSet.dependsOn(webTest)
             }
         }
     }
