+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/NavigatorProviderTest.kt
@@ -16,39 +16,14 @@
 
 package androidx.navigation
 
-import android.os.Bundle
+import androidx.core.bundle.Bundle
+import androidx.kruth.assertThat
+import androidx.kruth.assertWithMessage
 import androidx.navigation.testing.TestNavigatorState
-import com.google.common.truth.Truth.assertThat
-import com.google.common.truth.Truth.assertWithMessage
-import org.junit.Assert.fail
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
+import kotlin.test.Test
+import kotlin.test.fail
 
-@RunWith(JUnit4::class)
 class NavigatorProviderTest {
-    @Test
-    fun addWithMissingAnnotationName() {
-        val provider = NavigatorProvider()
-        val navigator = NoNameNavigator()
-        try {
-            provider.addNavigator(navigator)
-            fail(
-                "Adding a provider with no @Navigator.Name should cause an " +
-                    "IllegalArgumentException"
-            )
-        } catch (e: IllegalArgumentException) {
-            // Expected
-        }
-    }
-
-    @Test
-    fun addWithMissingAnnotationNameGetWithExplicitName() {
-        val provider = NavigatorProvider()
-        val navigator = NoNameNavigator()
-        provider.addNavigator("name", navigator)
-        assertThat(provider.getNavigator<NoNameNavigator>("name")).isEqualTo(navigator)
-    }
 
     @Test
     fun addWithExplicitNameGetWithExplicitName() {
@@ -58,35 +33,11 @@ class NavigatorProviderTest {
 
         assertThat(provider.getNavigator<EmptyNavigator>("name")).isEqualTo(navigator)
         try {
-            provider.getNavigator(EmptyNavigator::class.java)
+            provider.getNavigator<EmptyNavigator>(EmptyNavigator.NAME)
             fail("getNavigator(Class) with an invalid name should cause an IllegalStateException")
         } catch (e: IllegalStateException) {
             // Expected
         }
-    }
-
-    @Test
-    fun addWithExplicitNameGetWithMissingAnnotationName() {
-        val provider = NavigatorProvider()
-        val navigator = NoNameNavigator()
-        provider.addNavigator("name", navigator)
-        try {
-            provider.getNavigator(NoNameNavigator::class.java)
-            fail(
-                "getNavigator(Class) with no @Navigator.Name should cause an " +
-                    "IllegalArgumentException"
-            )
-        } catch (e: IllegalArgumentException) {
-            // Expected
-        }
-    }
-
-    @Test
-    fun addWithAnnotationNameGetWithAnnotationName() {
-        val provider = NavigatorProvider()
-        val navigator = EmptyNavigator()
-        provider.addNavigator(navigator)
-        assertThat(provider.getNavigator(EmptyNavigator::class.java)).isEqualTo(navigator)
     }
 
     @Test
@@ -160,49 +111,36 @@ class NavigatorProviderTest {
     }
 }
 
-class NoNameNavigator : Navigator<NavDestination>() {
-    override fun createDestination(): NavDestination {
-        throw IllegalStateException("createDestination is not supported")
-    }
+internal expect class NoNameNavigator() : Navigator<NavDestination> {
+    override fun createDestination(): NavDestination
 
     override fun navigate(
         destination: NavDestination,
         args: Bundle?,
         navOptions: NavOptions?,
         navigatorExtras: Extras?
-    ): NavDestination? {
-        throw IllegalStateException("navigate is not supported")
-    }
+    ): NavDestination?
 
-    override fun popBackStack(): Boolean {
-        throw IllegalStateException("popBackStack is not supported")
-    }
+    override fun popBackStack(): Boolean
 }
 
 /** An empty [Navigator] used to test [NavigatorProvider]. */
-@Navigator.Name(EmptyNavigator.NAME)
-internal open class EmptyNavigator : Navigator<NavDestination>() {
+internal expect open class EmptyNavigator() : Navigator<NavDestination> {
 
     companion object {
-        const val NAME = "empty"
+        val NAME: String
     }
 
-    override fun createDestination(): NavDestination {
-        throw IllegalStateException("createDestination is not supported")
-    }
+    override fun createDestination(): NavDestination
 
     override fun navigate(
         destination: NavDestination,
         args: Bundle?,
         navOptions: NavOptions?,
         navigatorExtras: Extras?
-    ): NavDestination? {
-        throw IllegalStateException("navigate is not supported")
-    }
+    ): NavDestination?
 
-    override fun popBackStack(): Boolean {
-        throw IllegalStateException("popBackStack is not supported")
-    }
+    override fun popBackStack(): Boolean
 }
 
 private const val NAME = "TEST"
