+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/serialization/NavArgumentGeneratorTest.kt
@@ -16,27 +16,22 @@
 
 package androidx.navigation.serialization
 
-import android.os.Bundle
-import android.os.Parcel
-import android.os.Parcelable
+import androidx.core.bundle.Bundle
+import androidx.kruth.assertThat
 import androidx.navigation.CollectionNavType
 import androidx.navigation.NamedNavArgument
 import androidx.navigation.NavArgument
 import androidx.navigation.NavType
 import androidx.navigation.navArgument
-import com.google.common.truth.Truth.assertThat
 import kotlin.reflect.typeOf
+import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.fail
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.serializer
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
 
-@RunWith(JUnit4::class)
 class NavArgumentGeneratorTest {
     @Test
     fun convertToInt() {
@@ -529,307 +524,6 @@ class NavArgumentGeneratorTest {
     }
 
     @Test
-    fun convertToParcelable() {
-        @Serializable
-        class TestParcelable : Parcelable {
-            override fun describeContents() = 0
-
-            override fun writeToParcel(dest: Parcel, flags: Int) {}
-        }
-
-        @Serializable class TestClass(val arg: TestParcelable)
-
-        val navType =
-            object : NavType<TestParcelable>(false) {
-                override fun put(bundle: Bundle, key: String, value: TestParcelable) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = TestParcelable()
-            }
-
-        val converted =
-            serializer<TestClass>().generateNavArguments(mapOf(typeOf<TestParcelable>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = false
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToParcelableNullable() {
-        @Serializable
-        class TestParcelable : Parcelable {
-            override fun describeContents() = 0
-
-            override fun writeToParcel(dest: Parcel, flags: Int) {}
-        }
-
-        @Serializable class TestClass(val arg: TestParcelable?)
-
-        val navType =
-            object : NavType<TestParcelable?>(true) {
-                override fun put(bundle: Bundle, key: String, value: TestParcelable?) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = TestParcelable()
-            }
-
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<TestParcelable?>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = true
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToParcelableArray() {
-        @Serializable
-        class TestParcelable : Parcelable {
-            override fun describeContents() = 0
-
-            override fun writeToParcel(dest: Parcel, flags: Int) {}
-        }
-
-        @Serializable class TestClass(val arg: Array<TestParcelable>)
-
-        val navType =
-            object : NavType<Array<TestParcelable>>(false) {
-                override fun put(bundle: Bundle, key: String, value: Array<TestParcelable>) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = emptyArray<TestParcelable>()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<Array<TestParcelable>>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = false
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToParcelableArrayNullable() {
-        @Serializable
-        class TestParcelable : Parcelable {
-            override fun describeContents() = 0
-
-            override fun writeToParcel(dest: Parcel, flags: Int) {}
-        }
-
-        @Serializable class TestClass(val arg: Array<TestParcelable>?)
-
-        val navType =
-            object : NavType<Array<TestParcelable>>(true) {
-                override fun put(bundle: Bundle, key: String, value: Array<TestParcelable>) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = emptyArray<TestParcelable>()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<Array<TestParcelable>?>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = true
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToSerializable() {
-        @Serializable class TestSerializable : java.io.Serializable
-
-        @Serializable class TestClass(val arg: TestSerializable)
-
-        val navType =
-            object : NavType<TestSerializable>(false) {
-                override fun put(bundle: Bundle, key: String, value: TestSerializable) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = TestSerializable()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<TestSerializable>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = false
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToSerializableNullable() {
-        @Serializable class TestSerializable : java.io.Serializable
-
-        @Serializable class TestClass(val arg: TestSerializable?)
-
-        val navType =
-            object : NavType<TestSerializable>(true) {
-                override fun put(bundle: Bundle, key: String, value: TestSerializable) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = TestSerializable()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<TestSerializable?>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = true
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToSerializableArray() {
-        @Serializable class TestSerializable : java.io.Serializable
-
-        @Serializable class TestClass(val arg: Array<TestSerializable>)
-
-        val navType =
-            object : NavType<Array<TestSerializable>>(false) {
-                override fun put(bundle: Bundle, key: String, value: Array<TestSerializable>) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = emptyArray<TestSerializable>()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<Array<TestSerializable>>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = false
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToSerializableArrayNullable() {
-        @Serializable class TestSerializable : java.io.Serializable
-
-        @Serializable class TestClass(val arg: Array<TestSerializable>?)
-
-        val navType =
-            object : NavType<Array<TestSerializable>>(true) {
-                override fun put(bundle: Bundle, key: String, value: Array<TestSerializable>) {}
-
-                override fun get(bundle: Bundle, key: String) = null
-
-                override fun parseValue(value: String) = emptyArray<TestSerializable>()
-            }
-        val converted =
-            serializer<TestClass>()
-                .generateNavArguments(mapOf(typeOf<Array<TestSerializable>?>() to navType))
-        val expected =
-            navArgument("arg") {
-                type = navType
-                nullable = true
-            }
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToEnum() {
-        @Serializable class TestClass(val arg: TestEnum)
-
-        val expected =
-            navArgument("arg") {
-                type = NavType.EnumType(TestEnum::class.java)
-                nullable = false
-            }
-        val converted = serializer<TestClass>().generateNavArguments()
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToTopLevelEnum() {
-        @Serializable class TestClass(val arg: TestTopLevelEnum)
-
-        val expected =
-            navArgument("arg") {
-                type = NavType.EnumType(TestTopLevelEnum::class.java)
-                nullable = false
-            }
-        val converted = serializer<TestClass>().generateNavArguments()
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToEnumNullable() {
-        @Serializable class TestClass(val arg: TestEnum?)
-
-        @Suppress("UNCHECKED_CAST")
-        val expected =
-            navArgument("arg") {
-                type = InternalNavType.EnumNullableType(TestEnum::class.java as Class<Enum<*>?>)
-                nullable = true
-            }
-        val converted = serializer<TestClass>().generateNavArguments()
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToNestedEnum() {
-        @Serializable class TestClass(val arg: EnumWrapper.NestedEnum)
-
-        val expected =
-            navArgument("arg") {
-                type = NavType.EnumType(EnumWrapper.NestedEnum::class.java)
-                nullable = false
-            }
-        val converted = serializer<TestClass>().generateNavArguments()
-        assertThat(converted).containsExactlyInOrder(expected)
-        assertThat(converted[0].argument.isDefaultValueUnknown).isFalse()
-    }
-
-    @Test
-    fun convertToEnumOverriddenSerialNameIllegal() {
-        @Serializable class TestClass(val arg: TestEnumCustomSerialName)
-
-        val exception =
-            assertFailsWith<IllegalArgumentException> {
-                serializer<TestClass>().generateNavArguments()
-            }
-        assertThat(exception.message)
-            .isEqualTo(
-                "Cannot find class with name \"MyCustomSerialName\". Ensure that the " +
-                    "serialName for this argument is the default fully qualified name"
-            )
-    }
-
-    @Test
     fun convertToEnumArray() {
         @Serializable class TestClass(val arg: Array<TestEnum>)
         val navType =
@@ -1307,65 +1001,48 @@ class NavArgumentGeneratorTest {
     // and hashcode which will need to be public api.
     private fun assertThat(actual: List<NamedNavArgument>) = actual
 
-    private fun List<NamedNavArgument>.containsExactlyInOrder(
-        vararg expectedArgs: NamedNavArgument
-    ) {
-        if (expectedArgs.size != this.size) {
-            fail("expected list has size ${expectedArgs.size} and actual list has size $size}")
-        }
-        for (i in indices) {
-            val actual = this[i]
-            val expected = expectedArgs[i]
-            if (expected.name != actual.name) {
-                fail("expected name ${expected.name}, was actually ${actual.name}")
-            }
+    enum class TestEnum {
+        TEST
+    }
+}
 
-            if (!expected.argument.isEqual(actual.argument)) {
-                fail(
-                    """expected ${expected.name} to be:
+internal fun List<NamedNavArgument>.containsExactlyInOrder(
+    vararg expectedArgs: NamedNavArgument
+) {
+    if (expectedArgs.size != this.size) {
+        fail("expected list has size ${expectedArgs.size} and actual list has size $size}")
+    }
+    for (i in indices) {
+        val actual = this[i]
+        val expected = expectedArgs[i]
+        if (expected.name != actual.name) {
+            fail("expected name ${expected.name}, was actually ${actual.name}")
+        }
+
+        if (!expected.argument.isEqual(actual.argument)) {
+            fail(
+                """expected ${expected.name} to be:
                 |   ${expected.argument}
                 |   but was:
                 |   ${actual.argument}
                 """
-                        .trimMargin()
-                )
-            }
-        }
-    }
-
-    private fun NavArgument.isEqual(other: NavArgument): Boolean {
-        if (this === other) return true
-        if (javaClass != other.javaClass) return false
-        if (isNullable != other.isNullable) return false
-        if (isDefaultValuePresent != other.isDefaultValuePresent) return false
-        if (isDefaultValueUnknown != other.isDefaultValueUnknown) return false
-        if (type != other.type) return false
-        // In context of serialization, we can only tell if defaultValue is present but don't know
-        // actual value, so we cannot compare it to the generated defaultValue. But if
-        // there is no defaultValue, we expect them both to be null.
-        return if (!isDefaultValuePresent) {
-            defaultValue == null && other.defaultValue == null
-        } else true
-    }
-
-    enum class TestEnum {
-        TEST
-    }
-
-    @SerialName("MyCustomSerialName")
-    enum class TestEnumCustomSerialName {
-        TEST
-    }
-
-    @Serializable
-    private class EnumWrapper {
-        enum class NestedEnum {
-            ONE,
-            TWO
+                    .trimMargin()
+            )
         }
     }
 }
 
-enum class TestTopLevelEnum {
-    TEST
+internal fun NavArgument.isEqual(other: NavArgument): Boolean {
+    if (this === other) return true
+    if (this::class != other::class) return false
+    if (isNullable != other.isNullable) return false
+    if (isDefaultValuePresent != other.isDefaultValuePresent) return false
+    if (isDefaultValueUnknown != other.isDefaultValueUnknown) return false
+    if (type != other.type) return false
+    // In context of serialization, we can only tell if defaultValue is present but don't know
+    // actual value, so we cannot compare it to the generated defaultValue. But if
+    // there is no defaultValue, we expect them both to be null.
+    return if (!isDefaultValuePresent) {
+        defaultValue == null && other.defaultValue == null
+    } else true
 }
