+++ navigation/navigation-common/src/commonTest/kotlin/androidx/navigation/serialization/NavTypeConverterTest.kt
@@ -16,10 +16,8 @@
 
 package androidx.navigation.serialization
 
-import android.os.Parcel
-import android.os.Parcelable
+import androidx.kruth.assertThat
 import androidx.navigation.NavType
-import com.google.common.truth.Truth.assertThat
 import kotlin.reflect.typeOf
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
@@ -32,10 +30,7 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.serializer
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
 
-@RunWith(JUnit4::class)
 class NavTypeConverterTest {
 
     @Test
@@ -469,20 +464,6 @@ class NavTypeConverterTest {
     }
 
     @Test
-    fun matchCustomParcelable() {
-        val descriptor = serializer<TestParcelable>().descriptor
-        val kType = typeOf<TestParcelable>()
-        assertThat(descriptor.matchKType(kType)).isTrue()
-    }
-
-    @Test
-    fun matchCustomSerializable() {
-        val descriptor = serializer<TestSerializable>().descriptor
-        val kType = typeOf<TestSerializable>()
-        assertThat(descriptor.matchKType(kType)).isTrue()
-    }
-
-    @Test
     fun matchCustomTypeNativeTypeParam() {
         @Serializable class TestClass<T : SerialDescriptor>
 
@@ -668,36 +649,6 @@ class NavTypeConverterTest {
     }
 
     @Test
-    fun getNavTypeParcelable() {
-        val type = serializer<TestParcelable>().descriptor.getNavType()
-        assertThat(type).isEqualTo(UNKNOWN)
-    }
-
-    @Test
-    fun getNavTypeParcelableArray() {
-        val type = serializer<Array<TestParcelable>>().descriptor.getNavType()
-        assertThat(type).isEqualTo(UNKNOWN)
-    }
-
-    @Test
-    fun getNavTypeSerializable() {
-        val type = serializer<TestSerializable>().descriptor.getNavType()
-        assertThat(type).isEqualTo(UNKNOWN)
-    }
-
-    @Test
-    fun getNavTypeSerializableArray() {
-        val type = serializer<Array<TestSerializable>>().descriptor.getNavType()
-        assertThat(type).isEqualTo(UNKNOWN)
-    }
-
-    @Test
-    fun getNavTypeEnumSerializable() {
-        val type = serializer<TestEnum>().descriptor.getNavType()
-        assertThat(type).isEqualTo(NavType.EnumType(TestEnum::class.java))
-    }
-
-    @Test
     fun getNavTypeEnumArraySerializable() {
         val type = serializer<Array<TestEnum>>().descriptor.getNavType()
         assertThat(type).isEqualTo(UNKNOWN)
@@ -756,15 +707,6 @@ class NavTypeConverterTest {
     @Serializable class ParamDerived : Param()
 
     @Serializable open class Param
-
-    @Serializable
-    class TestParcelable(val arg: Int, val arg2: String) : Parcelable {
-        override fun describeContents() = 0
-
-        override fun writeToParcel(dest: Parcel, flags: Int) {}
-    }
-
-    @Serializable class TestSerializable(val arg: Int, val arg2: String) : java.io.Serializable
 
     class ArgClass
