+++ navigation/navigation-common/src/jbTest/kotlin/androidx/navigation/NavigatorProviderTest.jb.kt
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.navigation
+
+import androidx.core.bundle.Bundle
+
+internal actual class NoNameNavigator
+actual constructor() : Navigator<NavDestination>() {
+    actual override fun createDestination(): NavDestination {
+        throw IllegalStateException("createDestination is not supported")
+    }
+
+    actual override fun navigate(
+        destination: NavDestination,
+        args: Bundle?,
+        navOptions: NavOptions?,
+        navigatorExtras: Extras?
+    ): NavDestination? {
+        throw IllegalStateException("navigate is not supported")
+    }
+
+    actual override fun popBackStack(): Boolean {
+        throw IllegalStateException("popBackStack is not supported")
+    }
+}
+
+/** An empty [Navigator] used to test [NavigatorProvider]. */
+internal actual open class EmptyNavigator
+actual constructor() : Navigator<NavDestination>(NAME) {
+
+    actual companion object {
+        actual const val NAME: String = "empty"
+    }
+
+    actual override fun createDestination(): NavDestination {
+        throw IllegalStateException("createDestination is not supported")
+    }
+
+    actual override fun navigate(
+        destination: NavDestination,
+        args: Bundle?,
+        navOptions: NavOptions?,
+        navigatorExtras: Extras?
+    ): NavDestination? {
+        throw IllegalStateException("navigate is not supported")
+    }
+
+    actual override fun popBackStack(): Boolean {
+        throw IllegalStateException("popBackStack is not supported")
+    }
+
+}
